Rename animation path
